Conflicts:
	bin/config.js
	bin/create.js
	bin/delete.js
	bin/info.js
	bin/update.js

+++ index.js
@@ -14,8 +14,9 @@ config.setCredentials = function (accessKeyId, secretAccessKey) {
 };
 
 // Run configuration wizard on a CFN template.
-config.configure = function(template, stackname, region, defaults, callback) {
-    inquirer.prompt(_(template.Parameters).map(_(config.question).partial(defaults)), function(answers) {
+config.configure = function(template, stackname, region, overrides, callback) {
+    var params = _(template.Parameters).map(_(config.question).partial(overrides));
+    inquirer.prompt(params, function(answers) {
         callback(null, {
             StackName: stackname,
             Region: region,
@@ -26,14 +27,17 @@ config.configure = function(template, stackname, region, defaults, callback) {
 
 // Return a inquirer-compatible question object for a given CFN template
 // parameter.
-config.question = function(defaults, parameter, key) {
+config.question = function(overrides, parameter, key) {
     var question = {
         name: key,
         message: key + '. ' + parameter.Description || key,
         filter: function(value) { return value.toString() }
     };
     if ('Default' in parameter) question.default = parameter.Default;
-    if (key in defaults) question.default = defaults[key];
+    if (key in overrides.defaults) question.default = overrides.defaults[key];
+    if (key in overrides.choices) question.choices = overrides.choices[key];
+    if (key in overrides.messages) question.message = overrides.messages[key];
+    if (key in overrides.filters) question.filter = overrides.filters[key];
 
     question.type = (function() {
         if (parameter.NoEcho === 'true') return 'password';
@@ -79,15 +83,10 @@ config.writeConfiguration = function(filepath, config, callback) {
     var filepath = path.resolve(path.join(filepath, config.StackName + '.cfn.json'));
     var json = JSON.stringify(config, null, 4);
 
-    console.log('About to write the following to %s:\n%s', filepath, json);
+    console.log('Stack configuration:\n%s', json);
 
-    inquirer.prompt([{
-        type: 'confirm',
-        name: 'confirm',
-        message: 'Is this ok?',
-        default: true
-    }], function(answers) {
-        if (!answers.confirm) return callback();
+    confirmAction('Okay to write this configuration to ' + filepath + '?', function(confirm) {
+        if (!confirm) return callback();
         fs.writeFile(filepath, json, callback);
     });
 };
@@ -100,10 +99,9 @@ config.writeConfiguration = function(filepath, config, callback) {
 // - name: Required. Name of the Cloudformation stack
 // - config: Optional. Path to a configuration file to use
 // - update: Defaults to false. Reads existing stack parameters.
-// - defaults: Defaults to {}. Can be overriden to provide your own defaults.
-//   Keys should be the parameter's name, values either a string or function
-//   If finding the default value is asychronous, then the funciton has to
-//   declare itself as such. See https://github.com/SBoudrias/Inquirer.js#question
+// - defaults, choices, messages, filters: Optional. Any of these properties can be
+//   set to an object where the keys are Cloudformation parameter names, and the
+//   values are as described by https://github.com/SBoudrias/Inquirer.js#question
 //
 //   Prioritization of defaults written by multiple processes follows:
 //   1. Values set by parameters in an existing Cloudformation stack
@@ -125,22 +123,32 @@ config.configStack = function(options, callback) {
             if (!options.update) return afterStackLoad(fileParameters, {});
             config.readStackParameters(options.name, options.region, function(err, stackParameters) {
                 if (err) return callback(err);
+
+                // Exclude masked stack parameters that come from the CFN API.
+                stackParameters = _(stackParameters).reject(function(param, key) {
+                    return template.Parameters[key].NoEcho === 'true';
+                });
+
                 afterStackLoad(fileParameters, stackParameters);
             });
         }
 
         function afterStackLoad(fileParameters, stackParameters) {
 
-            var defaults = _(stackParameters)
-                .chain()
-                .defaults(fileParameters)
-                .defaults(options.defaults)
-                .defaults(_(template.Parameters).reduce(function(memo, value, key) {
-                    memo[key] = value.Default;
-                    return memo;
-                }, {})).value();
+            var overrides = {
+                defaults: _(stackParameters).chain()
+                    .defaults(fileParameters)
+                    .defaults(options.defaults)
+                    .defaults(_(template.Parameters).reduce(function(memo, value, key) {
+                        memo[key] = value.Default;
+                        return memo;
+                    }, {})).value(),
+                choices: options.choices || {},
+                filters: options.filters || {},
+                messages: options.messages || {}
+            };
 
-            config.configure(template, options.name, options.region, defaults, function(err, configuration) {
+            config.configure(template, options.name, options.region, overrides, function(err, configuration) {
                 if (err) return callback(err);
                 config.writeConfiguration('', configuration, function(err, aborted) {
                     if (err) return callback(err);
@@ -167,17 +175,20 @@ config.createStack = function(options, callback) {
     config.configStack(options, function (err, configDetails) {
         if (err) return callback(err);
 
-        cfn.createStack({
-            StackName: options.name,
-            TemplateBody: JSON.stringify(configDetails.template, null, 4),
-            Parameters: _(configDetails.configuration.Parameters).map(function(value, key) {
-                return {
-                    ParameterKey: key,
-                    ParameterValue: value
-                };
-            }),
-            Capabilities: options.iam ? [ 'CAPABILITY_IAM' ] : []
-        }, callback);
+        confirmAction('Ready to create this stack?', function (confirm) {
+            if (!confirm) return callback();
+            cfn.createStack({
+                StackName: options.name,
+                TemplateBody: JSON.stringify(configDetails.template, null, 4),
+                Parameters: _(configDetails.configuration.Parameters).map(function(value, key) {
+                    return {
+                        ParameterKey: key,
+                        ParameterValue: value
+                    };
+                }),
+                Capabilities: options.iam ? [ 'CAPABILITY_IAM' ] : []
+            }, callback);
+        });
     });
 };
 
@@ -192,17 +203,20 @@ config.updateStack = function(options, callback) {
     config.configStack(options, function(err, configDetails) {
         if (err) return callback(err);
 
-        cfn.updateStack({
-            StackName: options.name,
-            TemplateBody: JSON.stringify(configDetails.template, null, 4),
-            Parameters: _(configDetails.configuration.Parameters).map(function(value, key) {
-                return {
-                    ParameterKey: key,
-                    ParameterValue: value
-                };
-            }),
-            Capabilities: options.iam ? [ 'CAPABILITY_IAM' ] : []
-        }, callback);
+        confirmAction('Ready to update the stack?', function (confirm) {
+            if (!confirm) return callback();
+            cfn.updateStack({
+                StackName: options.name,
+                TemplateBody: JSON.stringify(configDetails.template, null, 4),
+                Parameters: _(configDetails.configuration.Parameters).map(function(value, key) {
+                    return {
+                        ParameterKey: key,
+                        ParameterValue: value
+                    };
+                }),
+                Capabilities: options.iam ? [ 'CAPABILITY_IAM' ] : []
+            }, callback);
+        });
     });
 }
 
@@ -214,9 +228,12 @@ config.deleteStack = function(options, callback) {
         region: options.region
     }));
 
-    cfn.deleteStack({
-        StackName: options.name
-    }, callback);
+    confirmAction('Ready to delete the stack ' + options.name + '?', function (confirm) {
+        if (!confirm) return callback();
+        cfn.deleteStack({
+            StackName: options.name
+        }, callback);
+    })
 };
 
 config.stackInfo = function(options, callback) {
@@ -267,5 +284,16 @@ function readJsonFile(filelabel, filepath, callback) {
             return callback(e);
         }
         callback(null, jsonData);
+    });
+}
+
+function confirmAction(message, callback) {
+    inquirer.prompt([{
+        type: 'confirm',
+        name: 'confirm',
+        message: message,
+        default: true
+    }], function(answers) {
+        callback(answers.confirm);
     });
 }
